change package local->branch and split  DemoJourney
