+++ .ci/Jenkinsfile_cloud
@@ -11,6 +11,7 @@ pipeline {
     }
     parameters {
         string(name: 'STACK_VERSION', defaultValue: '7.x,8.0.0-SNAPSHOT', description: 'Stack version on Cloud')
+        string(name: 'SIMULATION', defaultValue: 'cloud.AtOnceJourney,cloud.DiscoverAtOnce,cloud.DemoJourney', description: 'Comma-separated simulation list')
         string(name: 'NUMBER_TEST_RUNS', defaultValue: '1', description: 'Running the same scenario up to 10 times')
         string(name: 'DEPLOY_SECRET', defaultValue: 'cloud-staging-api-key', description: 'Set Vault key for Cloud deployment')
     }
